% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{aggregated_nn_data_frame}
\alias{aggregated_nn_data_frame}
\alias{aggregated_nn_data_frame,ImageSet-method}
\alias{ANY,ANY-method}
\title{Dump the nearest neighbor data that was extracted}
\usage{
aggregated_nn_data_frame(x, ...)

\S4method{aggregated_nn_data_frame}{ImageSet}(x, use_pixel = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{use_pixel}{show the distances in pixels or micrometers (default: FALSE)}
}
\value{
data frame of aggrogated nn data
}
\description{
Dump the nearest neighbor data that was extracted
}
