% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R,
%   R/transform.functions.R
\docType{methods}
\name{as.data.frame,NNN-method}
\alias{as.data.frame,NNN-method}
\alias{as.data.frame,ImageSet-method}
\title{Get the dataframe from a normalized nearest neighbor object.}
\usage{
\S4method{as.data.frame}{NNN}(x)

\S4method{as.data.frame}{ImageSet}(x)
}
\arguments{
\item{x}{Normalized nearest neighbor object}

\item{...}{Additional arguments}

\item{x}{Iris ImageSet boject}

\item{...}{Additional arguments}
}
\value{
A dataframe

A dataframe
}
\description{
Get the dataframe from a normalized nearest neighbor object.

Read inForm output and store it in an IrisSpatialFeatures ImageSet object.
}
\examples{
dataset <- IrisSpatialFeatures_data
dataframe <- as.data.frame(dataset)
dataset <- IrisSpatialFeatures_data
dataframe <- as.data.frame(dataset)
}
