% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{binned_gated_nn_data_frame}
\alias{binned_gated_nn_data_frame}
\alias{binned_gated_nn_data_frame,ImageSet-method}
\alias{ANY,ANY-method}
\title{Extract and dump nearest neighbor raw data}
\usage{
binned_gated_nn_data_frame(x, ...)

\S4method{binned_gated_nn_data_frame}{ImageSet}(x, from, to, bin_width = 20,
  max_distance = 300, use_pixel = FALSE, split_frames = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{from}{MarkerA}

\item{to}{MarkerB}

\item{bin_width}{width of bin}

\item{max_distance}{fartherst distance from 0 to consider}

\item{use_pixel}{show the distances in pixels or micrometers (default: FALSE)}

\item{split_frames}{if true then output per-frame data instead of per-sample (default: FALSE)}
}
\value{
data frame of binned nn data
}
\description{
Extract and dump nearest neighbor raw data
}
