% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.functions.R
\docType{methods}
\name{collapse_markers}
\alias{collapse_markers}
\alias{collapse_markers,ImageSet-method}
\alias{collapse_markers,ANY,ANY-method}
\title{This function collapses two markers into one, and reruns the counting of cells.
Mostly a convenience function for the Shiny interface so we start with a completely split set
successively adding more markers}
\usage{
collapse_markers(image_set, marker1, marker2, combined)

\S4method{collapse_markers}{ImageSet}(image_set, marker1, marker2, combined)
}
\arguments{
\item{image_set}{IrisSpatialFeatures ImageSet object.}

\item{marker1}{Name of the first marker that should be collapsed.}

\item{marker2}{Name of the second marker that should be collapsed.}

\item{combined}{Name of the combined marker.}
}
\value{
IrisSpatialFeatures ImageSet object.
}
\description{
This function collapses two markers into one, and reruns the counting of cells.
Mostly a convenience function for the Shiny interface so we start with a completely split set
successively adding more markers
}
\examples{
dataset <- IrisSpatialFeatures_data
ds <- collapse_markers(dataset,marker1="SOX10+ PDL1+",marker2="SOX10+ PDL1-", combined="SOX10+")

}
