% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{methods}
\name{compare_normalized_nearest_neighbor}
\alias{compare_normalized_nearest_neighbor}
\alias{compare_normalized_nearest_neighbor,NNN,character,character,character-method}
\title{Compare distances between two markers to a reference marker.}
\usage{
compare_normalized_nearest_neighbor(NNN, markerA, markerB, reference,
  order = FALSE)


  \S4method{compare_normalized_nearest_neighbor}{NNN,character,character,character}(NNN,
  markerA, markerB, reference, order = FALSE)
}
\arguments{
\item{NNN}{Normalized nearest neighbor object}

\item{markerA}{Additional arguments}

\item{markerB}{Additional arguments}

\item{reference}{Additional arguments}

\item{order}{Optional character vector with sample names in an order for plotting}
}
\value{
Analysis data
}
\description{
Compare distances between two markers to a reference marker.
}
\examples{
dataset <- IrisSpatialFeatures_data
dataframe <- as.data.frame(dataset)
}
