% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Coordinate-class}
\alias{Coordinate-class}
\alias{Coordinate}
\title{Class to represent a single imaging coordinate}
\description{
Class to represent a single imaging coordinate
}
\section{Slots}{

\describe{
\item{\code{ppp}}{A spatstat ppp object that contains all coordinate information}

\item{\code{raw}}{Includes all raw data that is output from inForm}

\item{\code{mask}}{List of user defined masks that define regions of interest}

\item{\code{coordinate_name}}{Name of the current coordinate}

\item{\code{size_in_px}}{Size of the image in pixel, accounting for mask size}
}}

\examples{
coord <- new("Coordinate")

}
