% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.R
\docType{methods}
\name{extract_features}
\alias{extract_features}
\alias{extract_features,ImageSet-method}
\alias{extract_features,ANY,ANY-method}
\title{Extract all spatial features}
\usage{
extract_features(x, ...)

\S4method{extract_features}{ImageSet}(x, name = "", rm.na = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{name}{Prefix for all features, e.g. 'invasive_margin' (Default: '')}

\item{rm.na}{Should features with NA values be removed (Default: FALSE)}
}
\value{
dataframe of features
}
\description{
Extract all spatial features
}
\examples{

#' #loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
dataset <- extract_proximity(dataset,only_closest=TRUE,radii=25)
dataset <- extract_interactions(dataset)
extract_features(dataset)
}
