% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.functions.R
\name{extract_mask}
\alias{extract_mask}
\title{Read inForm output from a single coordinate}
\usage{
extract_mask(filename)
}
\arguments{
\item{filename}{Name of the .tif file that contains the mask.}
}
\value{
Mask matrix
}
\description{
Read inForm output from a single coordinate
}
\examples{
extract_mask(system.file("extdata",
                         "MEL2","MEL2_080416_1_Invasive_Margin.tif",
                         package = "IrisSpatialFeatures"))
}
