% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{extract_nearest_neighbor}
\alias{extract_nearest_neighbor}
\alias{extract_nearest_neighbor,ImageSet-method}
\alias{extract.nearest.neighbor,ANY,ANY-method}
\title{Extract the distance to each nearest neighbor for each cell-type}
\usage{
extract_nearest_neighbor(x, ...)

\S4method{extract_nearest_neighbor}{ImageSet}(x, min_num_cells = 10)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{min_num_cells}{Minimum number of cell that a coordinate needs to have in order to calculate the statistics (Default: 10)}
}
\value{
distance to nearest neighbor for each
}
\description{
Extract the distance to each nearest neighbor for each cell-type
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
extract_nearest_neighbor(dataset)

}
