% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.analysis.R
\docType{methods}
\name{extract_proximity}
\alias{extract_proximity}
\alias{extract_proximity,ImageSet-method}
\alias{run_proximity,ANY,ANY-method}
\title{Run a proximity analysis on all samples.
There are two modes this function can be run. In the first mode it uses the major and minor axes for each cell as provided by inform. It then averages
half of those two axes and adds an uncertainty margin, which are used to provide an estimate on whether two cells are touching. This mode can be used to approximate the interaction analysis.
The second mode uses a user specified distance to count the cells within the proximity of a given cell-type. With increasing distances usually cells fall into the proximity of multiple cells
of a given type so the function allows the restriction of only counting the cell only once.}
\usage{
extract_proximity(x, ...)

\S4method{extract_proximity}{ImageSet}(x, radii = c("Entire.Cell.Major.Axis",
  "Entire.Cell.Minor.Axis"), uncertainty_margin = 1, only_closest = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{radii}{The size of the radius in pixels or the default, which are the inForm output columns that indicated the minor and major axis of each cell.}

\item{uncertainty_margin}{Only for the approximation of the interaction analysis, where it indicates how many pixels further should be search to find a touching cell (deafult: 1).}

\item{only_closest}{For the proximity analysis a target cell can be in the vincinity of multiple source cells, so the counts are articicially inflated. E.g. a CD8 PD1+ T-cell is within <50 pixels of 30 HRS cells, this cell should only be counted for the closes HRS cell. (default: FALSE)}
}
\value{
The proximial events for each cell
}
\description{
Run a proximity analysis on all samples.
There are two modes this function can be run. In the first mode it uses the major and minor axes for each cell as provided by inform. It then averages
half of those two axes and adds an uncertainty margin, which are used to provide an estimate on whether two cells are touching. This mode can be used to approximate the interaction analysis.
The second mode uses a user specified distance to count the cells within the proximity of a given cell-type. With increasing distances usually cells fall into the proximity of multiple cells
of a given type so the function allows the restriction of only counting the cell only once.
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
extract_proximity(dataset)

}
