% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.functions.R
\docType{methods}
\name{extract_ROI}
\alias{extract_ROI}
\alias{extract_ROI,ImageSet-method}
\alias{extract_ROI,ANY,ANY-method}
\title{Method that reduces the current dataset to a specific region of interest, discarding all cell coordinates outside of that region}
\usage{
extract_ROI(x, ...)

\S4method{extract_ROI}{ImageSet}(x, ROI = "invasive_margin",
  verbose = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{ROI}{Region of interest (default: 'invasive_margin')}

\item{verbose}{Print some details (default: False)}
}
\value{
IrisSpatialFeatures ImageSet object
}
\description{
Method that reduces the current dataset to a specific region of interest, discarding all cell coordinates outside of that region
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
im_area <- extract_ROI(dataset,ROI='invasive_margin')

}
