% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Function to extract all numeric features}
\usage{
feature_selection(dat, lab)
}
\arguments{
\item{dat}{A data matrix with features as rows and samples as columns}

\item{lab}{Label annotation that contains 2 classes, which corresponds to the samples in the column}
}
\value{
t-test and wilcox test btween the 2 classes
}
\description{
Function to extract all numeric features
}
