% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{get_all_interactions}
\alias{get_all_interactions}
\alias{get_all_interactions,ImageSet-method}
\alias{get_all_interactions,ANY,ANY-method}
\title{Get all interactions between all cell-types}
\usage{
get_all_interactions(x, ...)

\S4method{get_all_interactions}{ImageSet}(x)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments.}
}
\value{
For each cell-type return interactions
}
\description{
Get all interactions between all cell-types
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(raw_data,
    marker='PD-Ligand-1 (Opal 690)',
    marker_name='PDL1',
    base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
    marker='PD-1 (Opal 540)',
    marker_name='PD1',
    base=c('CD8+','OTHER'))
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
dataset <- extract_proximity(dataset,only_closest=TRUE,radii=25)
dataset <- extract_interactions(dataset)
get_all_interactions(dataset)
}
