% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.analysis.R
\docType{methods}
\name{get_all_proximities}
\alias{get_all_proximities}
\alias{get_all_proximities,ImageSet-method}
\alias{get_all_proximities,ANY,ANY-method}
\title{Get all proximity data for all cell-types in a sample}
\usage{
get_all_proximities(x, ...)

\S4method{get_all_proximities}{ImageSet}(x)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
all proximity data for cell-types
}
\description{
Get all proximity data for all cell-types in a sample
}
\examples{
get_all_proximities(new("ImageSet"))

}
