% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.functions.R
\docType{methods}
\name{get_count_ratios}
\alias{get_count_ratios}
\alias{get_count_ratios,ImageSet-method}
\alias{get_count_ratios,ANY,ANY-method}
\title{Get ratio of counts between two markers}
\usage{
get_count_ratios(x, ...)

\S4method{get_count_ratios}{ImageSet}(x, marker1, marker2, digits = 2)
}
\arguments{
\item{x}{An IrisSpatialFeatures object}

\item{...}{Additional arguments.}

\item{marker1}{First cell-type.}

\item{marker2}{Second cell-type.}

\item{digits}{Number of digits that should be shown in the the results.
(Default: 2)}
}
\value{
Count ratio between two markers
}
\description{
Get ratio of counts between two markers
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
get_count_ratios(dataset,'SOX10+ PDL1-','SOX10+ PDL1+')
}
