% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.functions.R
\docType{methods}
\name{get_counts_per_mm2}
\alias{get_counts_per_mm2}
\alias{get_counts_per_mm2,ImageSet-method}
\title{Get all the counts on a per mm2 basis}
\usage{
get_counts_per_mm2(x, ...)

\S4method{get_counts_per_mm2}{ImageSet}(x, digits = 2, blank = FALSE)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{digits}{Number of digits that are shown in the output (default: 2)}

\item{blank}{(default: FALSE)}
}
\value{
counts per mm2 per sample, collapsing each coordinate and returning
        mean and standard error
}
\description{
Get all the counts on a per mm2 basis
}
\examples{
get_counts_per_mm2(new("ImageSet"))

}
