% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{get_interactions}
\alias{get_interactions}
\alias{get_interactions,ImageSet-method}
\alias{get_interactions,ANY,ANY-method}
\title{Get interactions for a specific marker}
\usage{
get_interactions(x, ...)

\S4method{get_interactions}{ImageSet}(x, marker, normalize = TRUE)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments.}

\item{marker}{Cell-type for which the interactions should be pulled}

\item{normalize}{Flag to indicated whether to normalize each sample so all interactions sum up to 1 (Default: 1)}
}
\value{
interactions for a specific marker
}
\description{
Get interactions for a specific marker
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(raw_data,
    marker='PD-Ligand-1 (Opal 690)',
    marker_name='PDL1',
    base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
    marker='PD-1 (Opal 540)',
    marker_name='PD1',
    base=c('CD8+','OTHER'))
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
dataset <- extract_proximity(dataset,only_closest=TRUE,radii=25)
dataset <- extract_interactions(dataset)
get_interactions(dataset,'CD8+ PD1+')
}
