% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{get_nearest_neighbors}
\alias{get_nearest_neighbors}
\alias{get_nearest_neighbors,ImageSet-method}
\alias{get_nearest_neighbors,ANY,ANY-method}
\title{Get the nearest neighbor for a specified cell-type}
\usage{
get_nearest_neighbors(x, ...)

\S4method{get_nearest_neighbors}{ImageSet}(x, marker)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{marker}{Cell type for which the nearest neighbor should be calculated}
}
\value{
nearest neighbors for the specified cell-type
}
\description{
Get the nearest neighbor for a specified cell-type
}
\examples{
#' #loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
get_nearest_neighbors(dataset,"SOX10+ PDL1+")

}
