% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.analysis.R, R/proximity.analysis.R
\docType{methods}
\name{get_proximities}
\alias{get_proximities}
\alias{get_proximities,ImageSet-method}
\alias{get_proximities,ANY,ANY-method}
\alias{get_proximities}
\alias{get_proximities,ImageSet-method}
\alias{get_proximities,ANY,ANY-method}
\title{Get proximity data for a given cell-type}
\usage{
get_proximities(x, ...)

\S4method{get_proximities}{ImageSet}(x, marker, normalize = TRUE)

get_proximities(x, ...)

\S4method{get_proximities}{ImageSet}(x, marker, normalize = TRUE)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments.}

\item{marker}{Cell type for which the proximity data should be extracted.}

\item{normalize}{Flag indicating whether the populations should be normalized so that the sum of all is 1 (default: TRUE).}

\item{x}{An IrisSpatialFeatures ImageSet object.}

\item{marker}{Cell type for which the proximity data should be extracted.}

\item{normalize}{Flag indicating whether the populations should be normalized so that the sum of all is 1 (default: TRUE).}

\item{...}{Additional arguments.}
}
\value{
proximities for a specific cell-type

proximities for a specific cell-type
}
\description{
Get proximity data for a given cell-type

Get proximity data for a given cell-type
}
