% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interaction_counts_sample_data_frame}
\alias{interaction_counts_sample_data_frame}
\alias{interaction_counts_sample_data_frame,ImageSet-method}
\alias{interaction_counts_sample_data_frame,ANY,ANY-method}
\title{DataFrame from all the interaction counts on a per mm2 basis per frame grouped by sample
This function is more like a density of edges between nodes on the interaction graph
than an average neighbor count}
\usage{
interaction_counts_sample_data_frame(x, verbose = FALSE, ...)

\S4method{interaction_counts_sample_data_frame}{ImageSet}(x, verbose)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments}
}
\value{
data frame
}
\description{
DataFrame from all the interaction counts on a per mm2 basis per frame grouped by sample
This function is more like a density of edges between nodes on the interaction graph
than an average neighbor count
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
interactions_sample_data_frame(dataset)

}
