% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interaction_maps}
\alias{interaction_maps}
\alias{interaction_maps,ImageSet-method}
\alias{interaction_maps,ANY,ANY-method}
\title{Plot interaction maps for all samples}
\usage{
interaction_maps(x, ...)

\S4method{interaction_maps}{ImageSet}(x, int_markers, int_marker_cols,
  silent_markers = c(), silent_col = c(), outline_transparency = 0.9,
  use_dapi = FALSE, outdir = "interaction_maps", useMask = NULL,
  format = ".png")
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments.}

\item{int_markers}{Cell-types that should be considered. If two cells
from different cell-types interact they are filled in,
if a cell is not interacting it is just outlined.}

\item{int_marker_cols}{Colors for the cell-types}

\item{silent_markers}{Cell-types that should only be outlined (Default: c())}

\item{silent_col}{Colors for silent markers (Default: c())}

\item{outline_transparency}{Dimming factor for the outlines cells(Default: 0.9)}

\item{use_dapi}{Use the DAPI channel as a background (Default: FALSE)}

\item{outdir}{Output directory (Default: './interaction_maps')}

\item{useMask}{(Default: NULL)}

\item{format}{Output format of the images. Can be '.png' or '.tiff' (Default: '.png')}
}
\value{
plot of interactions for all samples
}
\description{
Plot interaction maps for all samples
}
\examples{

#' #loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_interactions(dataset)
get_interactions(dataset,'CD8+ PD1+')
int_markers <- c('CD8+ PD1+','SOX10+ PDL1+')
int_marker_cols <- c('#dd1c77','#99d8c9')
silent_markers <- c('CD8+ PD1-')
silent_col=c('yellow')
p <- interaction_maps(dataset,int_markers,int_marker_cols,silent_markers,
                      silent_col)
}
