% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interaction_permutation_test}
\alias{interaction_permutation_test}
\alias{interaction_permutation_test,ImageSet-method}
\alias{interaction_permutation_test,ANY,ANY-method}
\title{Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption}
\usage{
interaction_permutation_test(x, permutations = 20, subset = NULL,
  verbose = FALSE, ...)

\S4method{interaction_permutation_test}{ImageSet}(x, permutations, subset,
  vebose)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object that has had extract nearest neighbors run}

\item{permutations}{Set to 100 by default}
}
\value{
data.frame of markers and distances
}
\description{
Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
interaction_permutation_test(dataset)

}
