% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interaction_proportion_comparison_permutation_test}
\alias{interaction_proportion_comparison_permutation_test}
\alias{interaction_proportion_comparison_permutation_test,ImageSet-method}
\alias{interaction_proportion_comparison_permutation_test,ANY,ANY-method}
\title{Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption}
\usage{
interaction_proportion_comparison_permutation_test(x, reference, phenotype_A,
  phenotype_B, permutations = 20, subset = NULL, verbose = FALSE, ...)

\S4method{interaction_proportion_comparison_permutation_test}{ImageSet}(x,
  reference, phenotype_A, phenotype_B, permutations, subset, verbose)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object that has had extract nearest neighbors run}

\item{reference}{Reference phenotype to compare proportions around}

\item{phenotype_A}{First cell type to compare proportions of}

\item{phenotype_B}{Second cell type to compare proportions of}

\item{permutations}{Set to 100 by default}

\item{subset}{Limit permutations to these types}
}
\value{
data.frame of markers and distances
}
\description{
Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
interaction_permutation_test(dataset)

}
