% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interaction_proportion_data_frame}
\alias{interaction_proportion_data_frame}
\alias{interaction_proportion_data_frame,ImageSet-method}
\alias{interaction_proportion_data_frame,ANY,ANY-method}
\title{DataFrame from all the interaction proportions freshly calculated from our current extracted counts}
\usage{
interaction_proportion_data_frame(x, verbose = FALSE, ...)

\S4method{interaction_proportion_data_frame}{ImageSet}(x, verbose)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments}
}
\value{
data frame
}
\description{
DataFrame from all the interaction proportions freshly calculated from our current extracted counts
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
interactions_sample_data_frame(dataset)

}
