% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{interactions_sample_data_frame}
\alias{interactions_sample_data_frame}
\alias{interactions_sample_data_frame,ImageSet-method}
\alias{interactions_sample_data_frame,ANY,ANY-method}
\title{DataFrame from all the counts on a per mm2 basis per sample}
\usage{
interactions_sample_data_frame(x, ...)

\S4method{interactions_sample_data_frame}{ImageSet}(x)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments}
}
\value{
data frame
}
\description{
DataFrame from all the counts on a per mm2 basis per sample
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
interactions_sample_data_frame(dataset)

}
