% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IrisSpatialFeatures.R
\docType{package}
\name{IrisSpatialFeatures}
\alias{IrisSpatialFeatures}
\alias{IrisSpatialFeatures-package}
\alias{IrisSpatialFeatures-package}
\title{IrisSpatialFeatures: A package to extract spatial features based on multiplex IF images}
\description{
Iris reads the output of the PerkinElmer inForm software and calculates
a variety of spatial statistics. In addition to simple counts, it can
derive average nearest neighbors for each cell-type and interaction
summary profiles for each celltype. These statistics are derived across
images, both overall and regions of interest as defined by user defined
masks.
}
\author{
Daniel Gusenleitner \email{Daniel_Gusenleitner@dfci.harvard.edu}
}
