% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IrisSpatialFeatures_data}
\alias{IrisSpatialFeatures_data}
\title{Example IrisSpatialFeatures dataset}
\format{IrisSpatialFeatures dataset}
\usage{
IrisSpatialFeatures_data
}
\description{
A dataset that contains two small samples, with 5 imaging coordinates per samples
}
\keyword{datasets}
