% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\docType{methods}
\name{launch_shiny_gui}
\alias{launch_shiny_gui}
\alias{launch_shiny_gui,ImageSet-method}
\alias{ANY,ANY-method}
\title{Launch a GUI to look at the data}
\usage{
launch_shiny_gui(x)

\S4method{launch_shiny_gui}{ImageSet}(x)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}
}
\value{
launched shiny server
}
\description{
Launch a GUI to look at the data
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
launch_shiny_gui(dataset)

}
