% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.analysis.R
\docType{methods}
\name{multiple_proximity_sample_data_frame}
\alias{multiple_proximity_sample_data_frame}
\alias{multiple_proximity_sample_data_frame,ImageSet-method}
\alias{multiple_proximity_sample_data_frame,ANY,ANY-method}
\title{Extract proximity for a range of distances.
Pass an array of radii distances (in micrometers) to the function 
that would otherwise work like "proximity_sample_data_frame"}
\usage{
multiple_proximity_sample_data_frame(x, ...)

\S4method{multiple_proximity_sample_data_frame}{ImageSet}(x, radii = c(10, 20,
  40, 100, 200, 400, 1000))
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{radii}{The array of distances to extract in microns.}
}
\value{
The proximial events for each cell
}
\description{
Extract proximity for a range of distances.
Pass an array of radii distances (in micrometers) to the function 
that would otherwise work like "proximity_sample_data_frame"
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
extract_proximity(dataset)

}
