% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{neighbor_ray_plot}
\alias{neighbor_ray_plot}
\alias{neighbor_ray_plot,ImageSet-method}
\alias{neighbor.ray.plot,ANY,ANY-method}
\title{Plot nearest neighbor ray plots for each samples}
\usage{
neighbor_ray_plot(x, ...)

\S4method{neighbor_ray_plot}{ImageSet}(x, from_type, to_type,
  from_col = "#EE7600", to_col = "#028482", format = ".pdf",
  plot_dir = "./", lineColor = "#666666", height = 7, width = 10)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments.}

\item{from_type}{Cell type from which the rays are drawn}

\item{to_type}{Cell type to which the rays are drawn}

\item{from_col}{Color for the 'from' cell-type (Default: '#EE7600')}

\item{to_col}{Color for the 'to' cell-type  (Default: '#028482')}

\item{format}{Format of the output file, can be '.pdf' or '.png' (Default: '.pdf')}

\item{plot_dir}{Directory in which the images are written (Default: './')}

\item{lineColor}{Color of the rays (Default: '#666666')}

\item{height}{Height of the pdf. (Default: 7)}

\item{width}{Width of the pdf. (Default: 10)}
}
\value{
nearest neighbor ray plots
}
\description{
Plot nearest neighbor ray plots for each samples
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset)
get_nearest_neighbors(dataset,"SOX10+ PDL1+")
plot_dir <- file.path('./ray_plots')
if (!file.exists(plot_dir)){
 dir.create(file.path(plot_dir))
}
}
