% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{methods}
\name{nn.test}
\alias{nn.test}
\alias{nn.test,ImageSet,character,character,character-method}
\alias{ANY,ANY-method}
\title{Calculate Nearest Neighbor for testing outputs}
\usage{
nn.test(x, reference, markerA, markerB)

\S4method{nn.test}{ImageSet,character,character,character}(x, reference,
  markerA, markerB)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object that has had extract nearest neighbors run}

\item{reference}{Reference marker}

\item{markerA}{First marker}

\item{markerB}{Second marker}
}
\value{
data.frame of markers and distances
}
\description{
Calculate Nearest Neighbor for testing outputs
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
extract_nearest_neighbor(dataset)

}
