% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{methods}
\name{nn_comparison_permutation_test}
\alias{nn_comparison_permutation_test}
\alias{nn_comparison_permutation_test,ImageSet,character,character,character,logical,numeric,character-method}
\alias{extract.nearest.neighbor,ANY,ANY-method}
\title{Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption}
\usage{
nn_comparison_permutation_test(x, markerA, markerB, reference,
  from_reference = TRUE, permutations = 20, subset = NULL)


  \S4method{nn_comparison_permutation_test}{ImageSet,character,character,character,logical,numeric,character}(x,
  markerA, markerB, reference, from_reference = TRUE, permutations = 20,
  subset = NULL)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object that has had extract nearest neighbors run}

\item{markerA}{First marker}

\item{markerB}{Second marker}

\item{reference}{Reference marker}

\item{from_reference}{If true calculate distance from the reference to the markers by NN}

\item{permutations}{Set to 100 by default}
}
\value{
data.frame of markers and distances
}
\description{
Calculate a permutation test result for nearest neighbors to say for each sample to see 
if the neighbor distance something seen under the null assumption
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
nn_comparison_permutation_test(dataset)

}
