% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{nn_data_frame}
\alias{nn_data_frame}
\alias{nn_data_frame,ImageSet-method}
\alias{ANY,ANY-method}
\title{Extract and dump nearest neighbor raw data}
\usage{
nn_data_frame(x, ...)

\S4method{nn_data_frame}{ImageSet}(x, from, to, use_pixel = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{from}{MarkerA}

\item{to}{MarkerB}

\item{use_pixel}{show the distances in pixels or micrometers (default: FALSE)}
}
\value{
data frame of aggrogated nn data
}
\description{
Extract and dump nearest neighbor raw data
}
