% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{class}
\name{NNN-class}
\alias{NNN-class}
\alias{NNN}
\title{Class to represent a normalized nearest neighbor.}
\description{
Class to represent a normalized nearest neighbor.
}
\section{Slots}{

\describe{
\item{\code{df}}{A dataframe of marker labels and nearest neighbor distances.}

\item{\code{micros_per_pixel}}{numeric for plotting scale}
}}

