% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{class}
\name{NNN_Compare-class}
\alias{NNN_Compare-class}
\alias{NNN_compare}
\title{Class to represent a comparison of two markers to a reference.}
\description{
Class to represent a comparison of two markers to a reference.
}
\section{Slots}{

\describe{
\item{\code{to_reference_plot}}{A plot comparing markerA and markerB's distance to the reference.}

\item{\code{to_reference_ttest}}{A paired ttest comparing markerA and markerB's distance to the reference.}

\item{\code{from_reference_plot}}{A plot comparing markerA and markerB's distance from the reference.}

\item{\code{from_reference_ttest}}{A ttest comparing markerA and markerB's distance from the reference.}

\item{\code{to_reference_order}}{For reordering other plots}

\item{\code{from_reference_order}}{For reordering other plots}

\item{\code{to_reference_df}}{The raw data}

\item{\code{from_reference_df}}{The raw data}
}}

