% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{methods}
\name{normal_nearest_neighbor_sample_once}
\alias{normal_nearest_neighbor_sample_once}
\alias{normal_nearest_neighbor_sample_once,character,ImageSet-method}
\alias{normal.nearest.neighbor.sample.once,ANY,ANY-method}
\title{Extract the distance to each nearest neighbor for specified
cell-types, normalized by downsampling each cell-type to the
same size, for a single sample, with no resampling}
\usage{
normal_nearest_neighbor_sample_once(sample_name, data, markers,
  minimum_cells = 50, grouped_sample = TRUE)

\S4method{normal_nearest_neighbor_sample_once}{character,ImageSet}(sample_name,
  data, markers, minimum_cells = 50, grouped_sample = TRUE)
}
\arguments{
\item{sample_name}{sample_name sample name string}

\item{data}{IrisSpatialFeatures ImageSet object}

\item{markers}{vector of marker names to use}

\item{minimum_cells}{the smallest number of cells (default:50)}

\item{grouped_sample}{TRUE/FALSE if we want to group samples together and
thus normalize the frames to the smallest frame
count (Default: TRUE)}
}
\value{
data.frame
}
\description{
Extract the distance to each nearest neighbor for specified
cell-types, normalized by downsampling each cell-type to the
same size, for a single sample, with no resampling
}
