% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\docType{methods}
\name{overview_plot}
\alias{overview_plot}
\alias{overview_plot,ImageSet-method}
\alias{overview_plot,ANY,ANY-method}
\title{Plot all coordinates in a given dataset}
\usage{
overview_plot(x, ...)

\S4method{overview_plot}{ImageSet}(x, outdir = "./", palette = NULL,
  type = "pdf", width = 10, height = 7)
}
\arguments{
\item{x}{Sample ImageSet object of the IrisSpatialFeatures package.}

\item{...}{Additional arguments}

\item{outdir}{Output directory (default: './')}

\item{palette}{Color palette used for the different cell-types. (default: NULL)}

\item{type}{File format for the plots. Can bei either 'pdf' or 'png'. (default: 'pdf')}

\item{width}{Width of the plot in inches for pdf and pixels for png. (default: 10)}

\item{height}{Heigth of the plot in inches for pdf and pixels for png. (default: 7)}
}
\value{
plot of all coordinates
}
\description{
Plot all coordinates in a given dataset
}
\examples{
overview_plot(new("ImageSet"))

}
