% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\docType{methods}
\name{plot,ImageSet-method}
\alias{plot,ImageSet-method}
\title{Plot all frames of an IrisSpatialFeatures ImageSet object.}
\usage{
\S4method{plot}{ImageSet}(x)
}
\arguments{
\item{x}{Iris ImageSet object}
}
\value{
A plot
}
\description{
Plot all frames of an IrisSpatialFeatures ImageSet object.
}
\examples{
dataset <- IrisSpatialFeatures_data
plot(dataset)
}
