% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{plot_nearest_neighbor}
\alias{plot_nearest_neighbor}
\alias{plot_nearest_neighbor,ImageSet-method}
\alias{plot_nearest_neighbor,ANY,ANY-method}
\title{Plot average nearest neighbor barplots for two cell types. This measurement is not symmetric, so if 'from' and 'to' are switched it will result in different results.
For the 'to' parameter this function allows a cell-type without '+' or '-' in the end. Indicating that the distances from the first cell-type should be calculated
against both '+/-' and a paired t-test should be calculated. For example we want to calculate the average distance between SOX10 PDL1+ melanoma cells against
both CD8 PD1+ and CD8 PD1- cells, the 'CD8 PD1' would be speficified as 'to' parameter, 2 distances would be calculated for each sample and a two-sided paired t-test calculated
to test for significant differences.}
\usage{
plot_nearest_neighbor(x, ...)

\S4method{plot_nearest_neighbor}{ImageSet}(x, from, to, ttest = TRUE,
  transposed = FALSE, remove_NAs = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments.}

\item{from}{Cell-type from which the nearest neighbor is calculated.}

\item{to}{Cell-type to which the nearest neighbor is calculated.}

\item{ttest}{Flag indicating whether a paired t-test should be calculated. (default: TRUE)}

\item{transposed}{Switches 'from' and 'to' cell-type. This way the (default: FALSE)}

\item{remove_NAs}{dont plot samples with less than min cells}
}
\value{
plot average nearest neighbor barplots for two cell types
}
\description{
Plot average nearest neighbor barplots for two cell types. This measurement is not symmetric, so if 'from' and 'to' are switched it will result in different results.
For the 'to' parameter this function allows a cell-type without '+' or '-' in the end. Indicating that the distances from the first cell-type should be calculated
against both '+/-' and a paired t-test should be calculated. For example we want to calculate the average distance between SOX10 PDL1+ melanoma cells against
both CD8 PD1+ and CD8 PD1- cells, the 'CD8 PD1' would be speficified as 'to' parameter, 2 distances would be calculated for each sample and a two-sided paired t-test calculated
to test for significant differences.
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset)
p <- plot_nearest_neighbor(dataset,'CD8+ PD1+','SOX10+ PDL1')

}
