% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.nearest.neighbor.R
\docType{methods}
\name{plot_nnn}
\alias{plot_nnn}
\alias{plot_nnn,tbl_df-method}
\title{Plot the normalized nearest neighbor (called internally)}
\usage{
plot_nnn(sub, markerA, markerB, reference, title, subtitle, legendtitle, order,
  font1, font2, microns_per_pixel)

\S4method{plot_nnn}{tbl_df}(sub, markerA, markerB, reference, title, subtitle,
  legendtitle, order = c(), font1 = 6, font2 = 4, microns_per_pixel = 1)
}
\arguments{
\item{sub}{subset of data in a tibble}

\item{markerA}{for factors}

\item{markerB}{for factors}

\item{reference}{for factors}

\item{title}{top of plot}

\item{subtitle}{on plot}

\item{legendtitle}{title to put over legend}

\item{order}{a vector of sample names to display from left to right,
if not specified will order by factor}

\item{font1}{font size 1}

\item{font2}{font size 2}

\item{microns_per_pixel}{scale data by this}
}
\value{
gg
}
\description{
Plot the normalized nearest neighbor (called internally)
}
