% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.analysis.R
\docType{methods}
\name{plot_proximities}
\alias{plot_proximities}
\alias{plot_proximities,ImageSet-method}
\alias{plot_proximities,ANY,ANY-method}
\title{Plot proximity analysis data}
\usage{
plot_proximities(x, ...)

\S4method{plot_proximities}{ImageSet}(x, label, ordering = NULL,
  normalize = TRUE, palette = NULL, celltype_order = NULL,
  xlim_fix = 13, topbar_cols = "darkgrey")
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional parameters.}

\item{label}{Cell-type for which the proximit profile is plotted}

\item{ordering}{Ordering of the samples (Default: NULL)}

\item{normalize}{Flag, should the populations of different cell-types sum up to one in each sample? (Default: TRUE)}

\item{palette}{Color palette, by default it uses Spectral from RColorbrewer (Default:NULL)}

\item{celltype_order}{Ordering of the cell-type. (Default: NULL)}

\item{xlim_fix}{Space on the right side to show the legend (Default: 13)}

\item{topbar_cols}{Color of the barplots on top (Default: 'darkgrey'}
}
\value{
plot proximity analysis
}
\description{
Plot proximity analysis data
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(raw_data,
    marker='PD-Ligand-1 (Opal 690)',
    marker_name='PDL1',
    base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
    marker='PD-1 (Opal 540)',
    marker_name='PD1',
    base=c('CD8+','OTHER'))
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
dataset <- extract_proximity(dataset,only_closest=TRUE,radii=25)
p <- plot_proximities(dataset,"SOX10+ PDL1-",xlim_fix=3)

}
