% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.analysis.R
\docType{methods}
\name{proximity_sample_data_frame}
\alias{proximity_sample_data_frame}
\title{Return a data frame of the per-sample proximities
like those output by the plot_proximties}
\usage{
proximity_sample_data_frame(x, ...)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
The proximial events for each cell
}
\description{
Return a data frame of the per-sample proximities
like those output by the plot_proximties
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
extract_proximity(dataset)

}
