% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{raw_data-class}
\alias{raw_data-class}
\alias{raw_data}
\title{Class to represent the raw data that is output by inForm}
\description{
Class to represent the raw data that is output by inForm
}
\section{Slots}{

\describe{
\item{\code{data}}{Raw coordinate data sheet}

\item{\code{summary}}{Inform summary (optional)}

\item{\code{score}}{Score table that indicates intensity cutoff for certain markers such as PD1 and PD-L1}

\item{\code{mem_seg_map}}{Membrane map as output by inForm}

\item{\code{nuc_seg_map}}{Nuclear map as output by inForm (optional)}

\item{\code{dapi_map}}{DAPI stain map as output by inForm (optional)}
}}

