% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{rayplot_single_coordinate}
\alias{rayplot_single_coordinate}
\alias{rayplot_single_coordinate,Coordinate-method}
\title{Plot nearest neighbor ray plots for a single coordinate}
\usage{
rayplot_single_coordinate(x, from_type, to_type, samp_name = "",
  from_col = "#EE7600", to_col = "#028482", lineColor = "#666666",
  use_pixel = FALSE, microns_per_pixel = 0.496)

\S4method{rayplot_single_coordinate}{Coordinate}(x, from_type, to_type,
  samp_name = "", from_col = "#EE7600", to_col = "#028482",
  lineColor = "#666666", use_pixel = FALSE, microns_per_pixel = 0.496)
}
\arguments{
\item{x}{An Coordinate object}

\item{from_type}{Cell type from which the rays are drawn}

\item{to_type}{Cell type to which the rays are drawn}

\item{samp_name}{Name of the sample}

\item{from_col}{Color for the 'from' cell-type (Default: '#EE7600')}

\item{to_col}{Color for the 'to' cell-type  (Default: '#028482')}

\item{lineColor}{Color for the line (Default: '#666666')}

\item{use_pixel}{Express units as pixels (Default: FALSE)}

\item{microns_per_pixel}{Conversion (Default: 0.496)}
}
\value{
a plot
}
\description{
Plot nearest neighbor ray plots for a single coordinate
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset)
rayplot_single_coordinate(x = dataset@samples[[1]]@coordinates[[1]],
                          samp_name = dataset@samples[[1]]@sample_name,
                          from_type = "SOX10+ PDL1+",
                          to_type = "CD8+ PD1+")
}
