% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Sample-class}
\alias{Sample-class}
\alias{Sample}
\title{An S3 class to represent a single imaging sample with multiple coordinates.}
\description{
An S3 class to represent a single imaging sample with multiple coordinates.
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{A list of coordinate objects that contain all of the raw and coordinate data.}

\item{\code{sample_name}}{Name of the contained sample.}
}}

