% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\docType{methods}
\name{shuffle_labels}
\alias{shuffle_labels}
\alias{shuffle_labels,ImageSet-method}
\alias{extract.nearest.neighbor,ANY,ANY-method}
\title{shuffle labels of each frame}
\usage{
shuffle_labels(x, ...)

\S4method{shuffle_labels}{ImageSet}(x, subset = NULL)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object that has had extract nearest neighbors run}
}
\value{
ImageSet with labels shuffled
}
\description{
shuffle labels of each frame
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
shuffle_labels(dataset)

}
