% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.functions.R
\docType{methods}
\name{threshold_dataset}
\alias{threshold_dataset}
\alias{threshold_dataset,ImageSet,character,character-method}
\alias{threshold_dataset,ANY,ANY-method}
\title{This function reads the manually determined thresholds of certain markers (e.g. PD1, PD-L1) and splits selected celltypes into marker+ and marker- celltypes.}
\usage{
threshold_dataset(image_set, marker, marker_name, ...)

\S4method{threshold_dataset}{ImageSet,character,character}(image_set, marker,
  marker_name, base = NULL, pheno_name = "Phenotype",
  remove_blanks = TRUE)
}
\arguments{
\item{image_set}{IrisSpatialFeatures ImageSet object.}

\item{marker}{Name of the marker used in the score file.}

\item{marker_name}{corresponding name, which should be appended at the selected cell types.}

\item{base}{Vector of cell types for which the marker should be used.}

\item{pheno_name}{Name of the phenotype column to be used. (Default from inForm is "Phenotype")}

\item{remove_blanks}{Flag that indicates whether or not not called cells are to be removed. (Default: TRUE)}
}
\value{
IrisSpatialFeatures ImageSet object.
}
\description{
This function reads the manually determined thresholds of certain markers (e.g. PD1, PD-L1) and splits selected celltypes into marker+ and marker- celltypes.
}
\examples{
dataset <- read_raw(path=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(dataset,
                             marker='PD-Ligand-1 (Opal 690)',
                             marker_name='PDL1',
                             base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
                             marker='PD-1 (Opal 540)',
                             marker_name='PD1',
                             base=c('CD8+','OTHER'))
}
