% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{extract_interactions}
\alias{extract_interactions}
\alias{extract_interactions,ImageSet-method}
\alias{extract_interactions,ANY,ANY-method}
\title{Extract interactions between all cell-types}
\usage{
extract_interactions(x, ...)

\S4method{extract_interactions}{ImageSet}(x)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
list of interactions
}
\description{
Extract interactions between all cell-types
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_interactions(dataset)

}
