% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{extract_interactions}
\alias{extract_interactions}
\alias{extract_interactions,ImageSet-method}
\alias{extract_interactions,ANY,ANY-method}
\title{Extract interactions between all cell-types}
\usage{
extract_interactions(x, ...)

\S4method{extract_interactions}{ImageSet}(x)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
list of interactions
}
\description{
Extract interactions between all cell-types
}
\examples{
extract_interactions(new("ImageSet"))

raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(raw_data,
    marker='PD-Ligand-1 (Opal 690)',
    marker_name='PDL1',
    base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
    marker='PD-1 (Opal 540)',
    marker_name='PD1',
    base=c('CD8+','OTHER'))
dataset <- extract_nearest_neighbor(dataset,min_num_cells=2)
dataset <- extract_proximity(dataset,only_closest=TRUE,radii=25)
dataset <- extract_interactions(dataset)

}
