% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.functions.R
\docType{methods}
\name{extract_ROI}
\alias{extract_ROI}
\alias{extract_ROI,ImageSet-method}
\alias{extract_ROI,ANY,ANY-method}
\title{Method that reduces the current dataset to a specific region of interest, discarding all cell coordinates outside of that region}
\usage{
extract_ROI(x, ...)

\S4method{extract_ROI}{ImageSet}(x, ROI = "invasive_margin")
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{ROI}{Region of interest (default: 'invasive_margin')}
}
\value{
IrisSpatialFeatures ImageSet object
}
\description{
Method that reduces the current dataset to a specific region of interest, discarding all cell coordinates outside of that region
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
extract_ROI(raw_data)
dataset <- threshold_dataset(raw_data,
                             marker='PD-Ligand-1 (Opal 690)',
                             marker_name='PDL1',
                             base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
                             marker='PD-1 (Opal 540)',
                             marker_name='PD1',
                             base=c('CD8+','OTHER'))                     
}
