% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{get_all_interactions}
\alias{get_all_interactions}
\alias{get_all_interactions,ImageSet-method}
\alias{get_all_interactions,ANY,ANY-method}
\title{Get all interactions between all cell-types}
\usage{
get_all_interactions(x, ...)

\S4method{get_all_interactions}{ImageSet}(x)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments.}
}
\value{
For each cell-type return interactions
}
\description{
Get all interactions between all cell-types
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_interactions(dataset)
get_all_interactions(dataset)

}
