% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{get_all_nearest_neighbors}
\alias{get_all_nearest_neighbors}
\alias{get_all_nearest_neighbors,ImageSet-method}
\alias{get_all_nearest_neighbors,ANY,ANY-method}
\title{Get the nearest neighbor for each cell-type}
\usage{
get_all_nearest_neighbors(x, ...)

\S4method{get_all_nearest_neighbors}{ImageSet}(x)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
Nearest neighbor for each cell-type
}
\description{
Get the nearest neighbor for each cell-type
}
\examples{
get_all_nearest_neighbors(new("ImageSet"))

}
