% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.analysis.R, R/proximity.analysis.R
\docType{methods}
\name{get_all_proximities}
\alias{get_all_proximities}
\alias{get_all_proximities,ImageSet-method}
\alias{get_all_proximities,ANY,ANY-method}
\alias{get_all_proximities}
\alias{get_all_proximities,ImageSet-method}
\alias{get_all_proximities,ANY,ANY-method}
\title{Get all proximity data for all cell-types in a sample}
\usage{
get_all_proximities(x, ...)

\S4method{get_all_proximities}{ImageSet}(x)

get_all_proximities(x, ...)

\S4method{get_all_proximities}{ImageSet}(x)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}

\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments}
}
\value{
all proximity data for cell-types

all proximity data for cell-types
}
\description{
Get all proximity data for all cell-types in a sample

Get all proximity data for all cell-types in a sample
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
get_all_proximities(dataset)


#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
get_all_proximities(dataset)

}
