% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.functions.R
\docType{methods}
\name{get_count_ratios}
\alias{get_count_ratios}
\alias{get_count_ratios,ImageSet-method}
\alias{get_count_ratios,ANY,ANY-method}
\title{Get ratio of counts between two markers}
\usage{
get_count_ratios(x, ...)

\S4method{get_count_ratios}{ImageSet}(x, marker1, marker2, digits = 2)
}
\arguments{
\item{x}{An IrisSpatialFeatures object}

\item{...}{Additional arguments.}

\item{marker1}{First cell-type.}

\item{marker2}{Second cell-type.}

\item{digits}{Number of digits that should be shown in the the results. (Default: 2)}
}
\value{
Count ratio between two markers
}
\description{
Get ratio of counts between two markers
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(raw_data,
                     raw_dir_name=system.file("extdata",
                             package = "IrisSpatialFeatures"),
                     format='Mantra')
dataset <- threshold_dataset(raw_data,
    marker='PD-Ligand-1 (Opal 690)',
    marker_name='PDL1',
    base=c('SOX10+'))
dataset <- threshold_dataset(dataset,
    marker='PD-1 (Opal 540)',
    marker_name='PD1',
    base=c('CD8+','OTHER'))
    get_count_ratios(dataset,'SOX10+ PDL1-','SOX10+ PDL1+')
}
