% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.analysis.R
\docType{methods}
\name{get_interactions}
\alias{get_interactions}
\alias{get_interactions,ImageSet-method}
\alias{get_interactions,ANY,ANY-method}
\title{Get interactions for a specific marker}
\usage{
get_interactions(x, ...)

\S4method{get_interactions}{ImageSet}(x, marker, normalize = TRUE)
}
\arguments{
\item{x}{An IrisSpatialFeatures ImageSet object}

\item{...}{Additional arguments.}

\item{marker}{Cell-type for which the interactions should be pulled}

\item{normalize}{Flag to indicated whether to normalize each sample so all interactions sum up to 1 (Default: 1)}
}
\value{
interactions for a specific marker
}
\description{
Get interactions for a specific marker
}
\examples{

#' #loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_interactions(dataset)
get_interactions(dataset,'CD8+ PD1+')

}
