% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbor.R
\docType{methods}
\name{plot_nearest_neighbor}
\alias{plot_nearest_neighbor}
\alias{plot_nearest_neighbor,ImageSet-method}
\alias{plot_nearest_neighbor,ANY,ANY-method}
\title{Plot average nearest neighbor barplots for two cell types. This measurement is not symmetric, so if 'from' and 'to' are switched it will result in different results.
For the 'to' parameter this function allows a cell-type without '+' or '-' in the end. Indicating that the distances from the first cell-type should be calculated
against both '+/-' and a paired sign rank should be calculated. For example we want to calculate the average distance between SOX10 PDL1+ melanoma cells against
both CD8 PD1+ and CD8 PD1- cells, the 'CD8 PD1' would be speficified as 'to' parameter, 2 distances would be calculated for each sample and a two-sided paired signed rank test calculated
to test for significant differences.}
\usage{
plot_nearest_neighbor(x, ...)

\S4method{plot_nearest_neighbor}{ImageSet}(x, from, to, signedrank = TRUE,
  transposed = FALSE, remove_NAs = FALSE, use_pixel = FALSE)
}
\arguments{
\item{x}{IrisSpatialFeatures ImageSet object.}

\item{...}{Additional arguments.}

\item{from}{Cell-type from which the nearest neighbor is calculated.}

\item{to}{Cell-type to which the nearest neighbor is calculated.}

\item{signedrank}{Flag indicating whether a paired signed rank test should be calculated. (default: TRUE)}

\item{transposed}{Switches 'from' and 'to' cell-type. This way the (default: FALSE)}

\item{remove_NAs}{dont plot samples with less than min cells}

\item{use_pixel}{show the distances in pixels or micrometers (default: FALSE)}
}
\value{
plot average nearest neighbor barplots for two cell types
}
\description{
Plot average nearest neighbor barplots for two cell types. This measurement is not symmetric, so if 'from' and 'to' are switched it will result in different results.
For the 'to' parameter this function allows a cell-type without '+' or '-' in the end. Indicating that the distances from the first cell-type should be calculated
against both '+/-' and a paired sign rank should be calculated. For example we want to calculate the average distance between SOX10 PDL1+ melanoma cells against
both CD8 PD1+ and CD8 PD1- cells, the 'CD8 PD1' would be speficified as 'to' parameter, 2 distances would be calculated for each sample and a two-sided paired signed rank test calculated
to test for significant differences.
}
\examples{

#loading pre-read dataset
dataset <- IrisSpatialFeatures_data
dataset <- extract_nearest_neighbor(dataset)
p <- plot_nearest_neighbor(dataset,'CD8+ PD1+','SOX10+ PDL1')

}
