% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.functions.R
\docType{methods}
\name{read_raw}
\alias{read_raw}
\alias{read_raw,character-method}
\alias{read_raw,ANY,ANY-method}
\title{Read inForm output and store it in an IrisSpatialFeatures ImageSet object.}
\usage{
read_raw(path, ...)

\S4method{read_raw}{character}(path, label_fix = list(), format = "Vectra",
  dir_filter = "", read_nuc_seg_map = FALSE, read_dapi_map = FALSE,
  use_binary_seg_maps = FALSE, MicronsPerPixel = 0.496,
  invasive_margin_in_px = 100, readMasks = TRUE, ROI = FALSE,
  ignore_scoring = FALSE)
}
\arguments{
\item{path}{Directory that contains the raw files}

\item{...}{Additional arguments}

\item{label_fix}{List of length 2 character vector that is used to fix filenames.}

\item{format}{Output format: Currently only "Vectra" and "Mantra" are supported.}

\item{dir_filter}{Filter to select only certain directory names.}

\item{read_nuc_seg_map}{Flag indicating whether the nuclear map should be read.}

\item{read_dapi_map}{Flag indicating whether the dapi layer should be read.}

\item{use_binary_seg_maps}{Flag indicating whether the dapi layer should be read.}

\item{MicronsPerPixel}{Length of one pixel. Default: 0.496, corresponding to a 20x Mantra/Vectra images}

\item{invasive_margin_in_px}{The width of the invasive margin in pixels}

\item{readMasks}{Flag indicating whether the "_Tumor.tif" and "_Invasive_Margin.tif" should be read (default: True)}

\item{ROI}{Flag indicating whether ROI was specified so the image is only analyzed within that region a '_ROI.tif' needs to be present. (default: False)}

\item{ignore_scoring}{Flag indicating whether the scoring file should be ignored (default: False)}
}
\value{
IrisSpatialFeatures ImageSet object.
}
\description{
Read inForm output and store it in an IrisSpatialFeatures ImageSet object.
}
\examples{
raw_data <- new("ImageSet")
raw_data <- read_raw(path=system.file("extdata", package = "IrisSpatialFeatures"),
                     format='Mantra')
}
